fix: lint error for new configuration
