+++ workspaces/app/src/App.tsx
@@ -1,12 +1,14 @@
 import { BrowserRouter } from "react-router-dom";
-import { LoadingOverlay } from "./components/loading-overlay/loading-overlay";
 import { AppRoutes } from "./routes";
 import { Suspense } from "react";
-import { LoadingOverlayProvider } from "./providers/loading-overlay";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { RepositoryProvider } from "./providers/repository";
 import { useSetup } from "./hooks/setup";
 import { ToastProvider } from "./components/ui/toast/toast-provider";
+import {
+  LoadingOverlay,
+  LoadingOverlayProvider,
+} from "./components/ui/loading-overlay";
 
 const AppRoot = () => {
   useSetup();
