--- workspaces/app/src/components/loading-overlay/index.ts
@@ -1,1 +1,0 @@
-export * from "./loading-overlay";
