+++ workspaces/app/src/components/ui/loading-overlay/loading-overlay-context.tsx
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+interface LoadingOverlayContextProps {
+  isLoading: boolean;
+  setIsLoading: (isLoading: boolean) => void;
+}
+
+export const LoadingOverlayContext = createContext<LoadingOverlayContextProps>({
+  isLoading: false,
+  setIsLoading: () => void 0,
+});
