+++ workspaces/app/src/components/ui/toast/index.ts
@@ -1,1 +1,7 @@
-export { Toast } from "./toast";
+import { ToastItem } from "./toast-item";
+import { ToastStack } from "./toast-stack";
+
+export const Toast = {
+  Item: ToastItem,
+  Stack: ToastStack,
+};
