+++ workspaces/app/src/components/ui/toast/toast-item.tsx
@@ -7,7 +7,7 @@ interface ToastItemProps {
   to?: string;
 }
 
-const ToastItem = ({ children, to }: ToastItemProps) => {
+export const ToastItem = ({ children, to }: ToastItemProps) => {
   if (to) {
     return (
       <Link to={to} className={clsx(styles.toastItemContainer, styles.link)}>
@@ -17,17 +17,4 @@ const ToastItem = ({ children, to }: ToastItemProps) => {
   }
 
   return <div className={styles.toastItemContainer}>{children}</div>;
-};
-
-interface ToastStackProps {
-  children: React.ReactNode;
-}
-
-const ToastStack = ({ children }: ToastStackProps) => {
-  return <div className={styles.toastStackContainer}>{children}</div>;
-};
-
-export const Toast = {
-  Item: ToastItem,
-  Stack: ToastStack,
 };
