+++ workspaces/app/src/components/ui/toast/toast-stack.tsx
@@ -1,0 +1,9 @@
+import { styles } from "./toast.css";
+
+interface ToastStackProps {
+  children: React.ReactNode;
+}
+
+export const ToastStack = ({ children }: ToastStackProps) => {
+  return <div className={styles.toastStackContainer}>{children}</div>;
+};
