+++ workspaces/app/src/components/ui/toast/use-toast.ts
@@ -1,5 +1,5 @@
-import { ToastContext } from "./toast-provider";
 import { useContext } from "react";
+import { ToastContext } from "./toast-context";
 
 export const useToast = () => {
   return useContext(ToastContext);
