+++ workspaces/app/src/hooks/setup.tsx
@@ -1,7 +1,7 @@
-import { LoadingOverlayContext } from "@/providers/loading-overlay";
 import { useContext, useEffect } from "react";
 import { useAuthUser } from "./auth-user";
 import { useNotifications } from "@/features/notification/use-notifications";
+import { LoadingOverlayContext } from "@/components/ui/loading-overlay";
 
 export const useSetup = () => {
   const { setIsLoading } = useContext(LoadingOverlayContext);
