+++ workspaces/app/src/libs/test-utils.tsx
@@ -33,7 +33,7 @@ interface CustomRenderOptions {
 const customRender = <
   Q extends typeof import("@testing-library/dom").queries,
   Container extends HTMLElement = HTMLElement,
-  BaseElement extends HTMLElement = HTMLElement
+  BaseElement extends HTMLElement = HTMLElement,
 >(
   ui: React.ReactElement,
   options: RenderOptions<Q, Container, BaseElement> & CustomRenderOptions = {}
@@ -50,7 +50,7 @@ const customRenderHook = <
   Props,
   Q extends typeof import("@testing-library/dom").queries,
   Container extends HTMLElement = HTMLElement,
-  BaseElement extends HTMLElement = HTMLElement
+  BaseElement extends HTMLElement = HTMLElement,
 >(
   render: (initialProps: Props) => Result,
   options: RenderOptions<Q, Container, BaseElement> & CustomRenderOptions = {}
@@ -62,6 +62,8 @@ const customRenderHook = <
   });
 };
 
+// 例外としてここはtest-utilityなので無視
+// eslint-disable-next-line react-refresh/only-export-components
 export * from "@testing-library/react";
 
 export { customRender as render, customRenderHook as renderHook };
