+++ workspaces/app/src/providers/repository/context.tsx
@@ -4,7 +4,7 @@ import { IChannelRepository } from "@/repositories/channel";
 import { IChatRepository } from "@/repositories/chat";
 import { IMessageRepository } from "@/repositories/message";
 import { INotificationRepository } from "@/repositories/notification";
-import React, { createContext } from "react";
+import { createContext } from "react";
 
 export interface RepositoryContextProps {
   channelRepository: IChannelRepository;
@@ -16,15 +16,3 @@ export interface RepositoryContextProps {
 
 export const RepositoryContext =
   createContext<RepositoryContextProps>(DefaultRepositories);
-
-interface LoadingProviderProps {
-  children: React.ReactNode;
-}
-
-export const RepositoryProvider = ({ children }: LoadingProviderProps) => {
-  return (
-    <RepositoryContext.Provider value={DefaultRepositories}>
-      {children}
-    </RepositoryContext.Provider>
-  );
-};
