+++ workspaces/app/src/providers/repository/provider.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { DefaultRepositories } from "@/repositories";
+import { RepositoryContext } from "./context";
+
+interface LoadingProviderProps {
+  children: React.ReactNode;
+}
+
+export const RepositoryProvider = ({ children }: LoadingProviderProps) => {
+  return (
+    <RepositoryContext.Provider value={DefaultRepositories}>
+      {children}
+    </RepositoryContext.Provider>
+  );
+};
